+++ CyRK/cy/cysolution.hpp
@@ -11,6 +11,7 @@
 #include "dense.hpp"
 #include "cysolver.hpp"
 #include "rk.hpp"
+#include "bdf.hpp"
 
 
 class CySolverResult {