+++ CyRK/cy/pysolver.pyx
@@ -79,11 +79,13 @@ cdef class PySolver(WrapCySolverResult):
             integration_method = ODEMethod.RK45
         elif method == 'dop853':
             integration_method = ODEMethod.DOP853
+        elif method == 'bdf':
+            integration_method = ODEMethod.BDF
         else:
             raise NotImplementedError(
                 "ERROR: `PySolver::set_problem_parameters` - "
                 f"Unknown or unsupported integration method provided: {method}.\n"
-                f"Supported methods are: RK23, RK45, DOP853."
+                f"Supported methods are: RK23, RK45, DOP853, BDF."
                 )
 
         cdef CySolverResult* cyresult_ptr = self.cyresult_uptr.get()